backend/requirements.txt
fastapi==0.115.6
google_search_results==2.4.2
pydantic==2.10.5
python-dotenv==1.0.1
Requests==2.32.3
serpapi==0.1.5
uvicorn==0.20.0
